[TogetherPay] refactor: 여행 연관관계 추가로 인한 Settlement 엔티티 수정

+++ src/main/java/com/devcrew/togetherpay/domain/settlement/Settlement.java
@@ -26,7 +26,7 @@ public class Settlement {
   public Long id;
 
   @Column(nullable = false)
-  private Integer amount;
+  private Long amount;
 
   @Builder.Default
   @Column(nullable = false)
@@ -40,7 +40,7 @@ public class Settlement {
   @JoinColumn(name = "user_id")
   User user;
 
-  public static Settlement of(Integer amount, Expense expense, User user) {
+  public static Settlement of(Long amount, Expense expense, User user) {
     return Settlement.builder()
         .amount(amount)
         .expense(expense)
